put in common the common files and code of thread and mutex str.c: fix a compile warning about LONG_MIN not fitting in int variable strklos.h: made EXTERN_PRIMITIVE actually declare the primitive extern, and  also declare its associated primitive_obj as extern. vm.c: fix (??) REG_CALL_PRIM which was doing some real wrong stuff I have no  idead why it worked, and if it did work, no idea why it still works after  my fix... all the rest: put the ';' at the end of EXTERN_PRIMITIVE(...) as it should  have been (or not?)

+++ src/str.c
@@ -90,7 +90,7 @@ static int stringcompi(SCM s1, SCM s2)
 static SCM control_index(int argc, SCM *argv, int *pstart, int *pend)
 {
   SCM s = NULL;
-  int len, start=0, end=-1;
+  long len, start=0, end=-1;
 
   /* Controling number of arguments */
   switch (argc) {
@@ -384,7 +384,7 @@ doc>
 
 DEFINE_PRIMITIVE("substring", substring, subr3, (SCM string, SCM start, SCM end))
 {
-  int from, to;
+  long from, to;
 
   if (!STRINGP(string)) error_bad_string(string);
